LayoutTests/http/tests/media/fairplay/legacy-fairplay-mse-v3-expected.txt


Created mediaSource
EVENT(sourceopen)
-
Appending Encrypted Video Header
FETCH: content/elementary-stream-video-header-keyid-1.m4v OK
EVENT(webkitneedkey)
EVENT(webkitkeymessage)
EXPECTED (uInt8ArrayToString(event.message) == 'certificate') OK
FETCH: resources/cert.der OK
EVENT(webkitkeymessage)
PROMISE: licenseResponse resolved
-
Appending Encrypted Video Payload
FETCH: content/elementary-stream-video-payload.m4v OK
EVENT(updateend)
EVENT(canplay)
END OF TEST
